+++ changelog.md
@@ -15,6 +15,7 @@
   + `discard-returns`: check whether the return value is discarded.
 * `NEW` locale `pt-br`, thanks [Jeferson Ferreira](https://github.com/jefersonf)
 * `NEW` supports [utf-8-offsets](https://clangd.llvm.org/extensions#utf-8-offsets)
+* `NEW` supports quickfix for `.luarc.json`
 * `CHG` `LuaDoc`:
   + `---@class` can be re-declared
   + supports unicode